requirements.txt
aioredis==1.2.0
asgiref==3.2.1
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
autopep8==1.4.4
bleach==3.1.0
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
Django==2.2.3
django-appconf==1.0.3
django-compressor==2.3
django-markdownify==0.8.0
django-sass-processor==0.7.3
djangorestframework==3.10.3
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.1
libsass==0.19.2
Markdown==3.1.1
mccabe==0.6.1
msgpack==0.6.1
Pillow==6.1.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pycparser==2.19
PyHamcrest==1.9.0
pylint==2.3.1
pytz==2019.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
typed-ast==1.4.0
webencodings==0.5.1
wrapt==1.11.2
zope.interface==4.6.0
